+++ test/azure/Expected/AcceptanceTests/Lro/lro/models/_models_py3.py
@@ -16,9 +16,9 @@ from ._auto_rest_long_running_operation_test_service_enums import *
 class OperationResult(msrest.serialization.Model):
     """OperationResult.
 
-    :param status: The status of the request. Possible values include: 'Succeeded', 'Failed',
-     'canceled', 'Accepted', 'Creating', 'Created', 'Updating', 'Updated', 'Deleting', 'Deleted',
-     'OK'.
+    :param status: The status of the request. Possible values include: "Succeeded", "Failed",
+     "canceled", "Accepted", "Creating", "Created", "Updating", "Updated", "Deleting", "Deleted",
+     "OK".
     :type status: str or ~lro.models.OperationResultStatus
     :param error:
     :type error: ~lro.models.OperationResultError
@@ -130,8 +130,8 @@ class Product(Resource):
     :vartype name: str
     :param provisioning_state:
     :type provisioning_state: str
-    :ivar provisioning_state_values:  Possible values include: 'Succeeded', 'Failed', 'canceled',
-     'Accepted', 'Creating', 'Created', 'Updating', 'Updated', 'Deleting', 'Deleted', 'OK'.
+    :ivar provisioning_state_values:  Possible values include: "Succeeded", "Failed", "canceled",
+     "Accepted", "Creating", "Created", "Updating", "Updated", "Deleting", "Deleted", "OK".
     :vartype provisioning_state_values: str or ~lro.models.ProductPropertiesProvisioningStateValues
     """
 
@@ -225,8 +225,8 @@ class SubProduct(SubResource):
     :vartype id: str
     :param provisioning_state:
     :type provisioning_state: str
-    :ivar provisioning_state_values:  Possible values include: 'Succeeded', 'Failed', 'canceled',
-     'Accepted', 'Creating', 'Created', 'Updating', 'Updated', 'Deleting', 'Deleted', 'OK'.
+    :ivar provisioning_state_values:  Possible values include: "Succeeded", "Failed", "canceled",
+     "Accepted", "Creating", "Created", "Updating", "Updated", "Deleting", "Deleted", "OK".
     :vartype provisioning_state_values: str or
      ~lro.models.SubProductPropertiesProvisioningStateValues
     """